modify: interface PlanType, Auth
