+++ src/types/interface.tsx
@@ -1,8 +1,9 @@
 import { InputType } from "@/common/components/Input";
-import auth, { Auth } from "@/common/redux/features/auth";
+import auth from "@/common/redux/features/auth";
 import { changeIDType } from "@/common/redux/features/changeID";
 import { BodyRateType } from "@/common/redux/features/dietary-record/bodyRate";
 import { GoalType } from "@/common/redux/features/dietary-record/goal";
+import { PaymentDataType } from "@/common/redux/features/paymentPhases";
 import { ShowModalType } from "@/common/redux/features/showModal";
 import { ReactNode } from "react";
 export interface LayoutProps {
@@ -32,6 +33,7 @@ export interface RegisterData {
 
 export interface RootState {
   registerPhases?: RegisterData | undefined;
+  paymentPhases: PaymentDataType;
   auth: Auth;
   showModal: ShowModalType;
   changeID: changeIDType;
@@ -81,4 +83,20 @@ export interface ComponentType {
   id?: string;
   options?: Array<{ option: string; value: string }>;
   disabled?: boolean;
+}
+
+export interface PlanType {
+  Title: string;
+  CourseName: string;
+  CourseWeek: string;
+  CoursePrice: string;
+}
+
+export interface Auth {
+  Token: string;
+  UserName: string;
+  Email: string;
+  ImgUrl: string;
+  IsNutritionist: boolean;
+  UserCurrentStatus: string;
 }
